src/requirements.txt
Django >=4.2.8, <5
Pillow
django-tinymce <3.7
django-nested-admin
flup6
django-fastcgi-server
django_sendmail_backend
